Uses Artifact Registry build images (#7)
* Uses Artifact Registry build images

And the name of the build image is stored as a user-defined substitution
variable in the build trigger in each project. This is because the flow for
this repo is mlab-sandbox -> mlab-staging -> measurement-lab, but the build
images are not present in the measurement-lab project. That project uses the
mlab-oti image, since both are production projects.

* Substitutes _only_ the image project

Rather than substituting the entire build image path, this commit changes that
to only substitute the project of the build image to make it more transparent
in the code what is being run.