Adds a small sleep before deploying dispatch.yaml
A previous build of this repo in the measurement-lab failed on the step trying
to deploy dispatch.yaml, and this was because the previous build step had
triggered some operations that were still in progress when the dispatch.yaml
step started, and app engine did not like this.

This change adds a 30s sleep before deploying dispatch.yaml just to be 100%
sure that all app engine operations that the previous step triggered have
completed. There is probably a more elegant solution using:

gcloud app operations wait

... but this would require basically a shell script as a build step, whereas a
simple sleep should work just fine. The operations which caused the previous
build failure completed just seconds after the dispatch.yaml had started, so
30s should be more than enough.