Moves all production deployment to a single build step
Previously, one build step would deploy app.yaml in production, and then a
second build step would deploy dispatch.yaml. This was causing build failures
on the 2nd step because even though the first build step completed as far as
Cloud Build was concerned, GCP had launched some other operations in the
background to promote the new deployment and deleted the old one. The second
step was failing because gcloud wouldn't launch a new operation for the service
while there were still operations in progress.

This change moves all production deployments steps to a single build step, and
even more to a single invocation of `gcloud app deploy`. In this way, gcloud
can decide when it it safe to deploy app.yaml and dispatch.yaml in relation to
one another.